issue #1 Create class SSchema to try traits Scala
The class instantiates a DatabaseConnection through an untyped instantiation of JdbcDatabaseConnectionManager (expressed as JdbcDatabaseConnectionManager[_] in Scala) that corresponds in Java to JdbcDatabaseConnectionManager<?>, which javac don't accept:

    [javac] Compiling 1 source file to /home/ivo/project/dbtune/build/classes
    [javac] extensions/Console/src/edu/ucsc/dbtune/core/Schema.java:129: incompatible types; inferred type argument(s) java.lang.Object do not conform to bounds of type variable(s) I
    [javac] found   : <I>edu.ucsc.dbtune.core.DatabaseConnectionManager<I>
    [javac] required: java.lang.Object
    [javac]         return new Schema(makeDatabaseConnectionManager(props).connect());
    [javac]                                                        ^

class Schema is just an example of how to call a Scala object from Java. In this case, Schema is useless since all the main functionality is already implemented in SSchema, but it still serves as example of how to interop between Java and Scala.