issue #31 Add sample 'movies' test to be used for tests
The script resources/test/postgres/movies.sql can be used to load the 'movies' sample database.

The logic for testing a database was added to ant through the build.properties and the tasks.xml
file. the build.properties file contains the configuration of the database under test, while the
tasks.xml was extended to pass these connection properties down to the JVM environment.

Given a java.sql.Connection object, the util.SQLScriptExecuter class executes a set of statements
contained in a file.