closes #54 (again) Fix for functional and unit tests
The primitive ant tasks are:

 * compile
 * compile.tests
 * test.unit
 * test.functional

Functional tests should have suffix `TestFunctional` in its name, eg. `MetaDataExtractorTestFunctional.java`. From now on, unit and functional tests are completely separated. The composed tasks are:

 * compile.all
 * test.unit.all
 * test.functional.all
 * test.unit.and.functional
 * test.unit.and.functional.all

----------

added annotations to `package-info.java` files to avoid having ant recompile them everytime that the compile task is invoked (couldn't find a way to do it for scala)

this commit also contains a fix for bug in scala's `Configuration` and `Database` that was triggered by the renaming of JdbcDatabaseConnectionManager and DatabaseConnectionManager.