issue #68 Modified the way a vector is created.
Instead of considering the cardinality, we know consider the cost of each operator since it yields to better clusterings. The `PGOptimizer` was changed in order to obtain the per-operator cost since the costs contained in the EXPLAIN are accumulated costs, i.e. the cost that the plan takes up to that node.