fixes #84 internal IDs incorrectly handled in PGCommands
After carefully reviewing both our code and Karl's, the issue turned out to be in the way Karl's handles IDs in the particular case of the candidate extractor.

The candidates that are considered for recommendation are produced by the `RECOMMEND INDEXES` command. When this command is executed, the internal IDs are indexed from 1.

Karl's `IndexExtractor` takes the index list and stores them into a `CandidatePool` object. When indexes are added to the candidate pool, their IDs are reset and begin from 0. Since our version of the `IndexExtractor` produces an `Iterable<DBIndex>`, the indexes are never reset and thus the calculation of the `updateCost` value (in `PGCommands`) throws an `IndexOutOfBounds` exception. This is solved by indexing indexes in the `PGCommands` from 0 instead of 1.