issue #95 making PreparedSQLStatement what-if-optimizable
This change abstracts the concept of having a PreparedSQLStatement doing what-if optimization based on the information it holds.

This will enable the decoupling of IBG-dependent code

Still missing is the decoupling of the InteractionBank out of the IBGPreparedSQLStatement. This is not directly used by an IBGPreparedStatement so it should be straight-forward to remove from it