issue #92 unit testing of DatabaseSystem
Defined some static factory methods to ease the testing through mocking.

Also added powermock for mockito and its dependencies. PowerMock allows to test static classes and methods. This is great since the dev doesn't need to tweak the design in order to make it testable (which eventually leads to unnecessary complexity)

A couple of tests are missing in DatabaseSystemTest so that it can be considered "finished"

DbTuneMocks contains a static method to construct a simple mock of an Optimizer