issue #101 organization of unit/functional tests
This commit introduces a proposal for organizing unit/functional tests. The idea is to have code that is independent of a unit/functional test separated as static methods. This code checks the assertions that are defined for a given class, without caring about how the instances of the class under test are created. By organizing code in this way, we can have unit and functional tests reuse the same code ([DRY](http://en.wikipedia.org/wiki/DRY)).

Check the OptimizerTest and OptimizerFunctionalTest for an example of how to organize unit/functional tests based on the above proposal. In short, all the code that uses mocks goes into *Test.java. Then, the same *Test.java file should separate the mock-dependent code from the mock-independent into static functions that contain the assertions. Then, the functional test just redefines the tests but using "real" object instances and invokes that static methods by passing the "real" objects to them.

In simple words, *Test and *FunctionalTest are very similar, they just differ in the setUp() method: *Test sets the DatabaseSystem to use mocks, while *FunctionalTest communicates with an actual DatabaseSystem. Also *Test contains the generic (static) code, while *FunctionalTest takes it from there.

Also in this commit are contained skeletons of tests