issue #101 Adding tests to check update costs
This extends the tests to cover all the of the Optimizer functionality, it includes tests that check that costs of updated indexes are retrieved correctly.

This commit completes the basic test coverage for the optimizer interface. After mocks are written for the Optimizer dependencies, the unit test will be completed (just need to uncomment the commented lines).