Modifying to ease maintenance
build.xml contains a note regarding the dependency between target names and the corresponding entry in the Wiki. If these names are modified, the Wiki entry should be modified accordingly.

EnvironmentProperties now contains a property that gives the list of vendors that are supported by the API (SUPPORTED_VENDORS). The same has been done with the list of supported optimizers (SUPPORTED_OPTIMIZERS).