fixes #109 Handling of fully qualified names
Removes the functionality implemented in implementations of
DatabaseObject and transfers it to the abstract class in a generic way
by defining abstract container and containees objects. The metadata
hierarchy is now traversed in a recursive way using this concept of
container and containees, while the particularities of each level have
been removed.

Also as part of the container/containee generalization, this commit
enables the dbtune/metadata package to receive fully qualified paths and
appropriately refer to corresponding database objects. For example:

   catalog.findByQualifiedName("schema_0.table_0.column_0")

returns a reference to the 'column_0' Column object. Another one:

   Index idx = catalog.<Index>create("schema_3.index_0");

creates an index called 'index_0' (using an abstract
newContainee(String) method) and adds it to 'schema_3'. This feature
will be used a lot by the CLI to hide the tedious Java-oriented syntax
from the user and provide a more SQL-like one.

Iterators at each level have been added. In order to traverse the
hierarchy:

   for(Schema schema : catalog.schemas()) {

       for(Table table : schema.tables()) {
           for(Column col : table.columns()) {
           }
       }

       for(Index index : schema.indexes()) {
           for(Column col : indexes.columns()) {
           }
       }
   }

These iterators allow the dropping of the Scala mirrors on the CLI
extension.

Convenience methods at each level are also preserved:

  Schema.getCatalog()
  Schema.findTable(String)
  Schema.findIndex(String)
  Table.getSchema()
  Table.findColumn(String)
  Index.getSchema()
  Index.getColumns()
  Column.getTable()

among others.