issue #124 Adding Index.inMemoryID()
This property is the equivalent of Karl's Index.internalID().

The Index.IN_MEMORY_ID is an atomic integer counter that always
increases as new indexes are created. This allows the easy mapping of
indexes to array-based structures (by keeping a consistent association
of index objects to the [0-MAX_INT] range).