issue #114 Modfile implementation of full tab scan
Modfy class IndexFullTableScan so that there is one index full table scan for each relation in each query
(i.e., if a relation R appears in two queries of the workload, then we have two index full table scans
objects on R).

Use getFullQualifiedName() of an index for the place that needs a unique ID of an index.