fixes #115 Operator information extraction
Adds two methods to DB2Optimizer (parsePlan and parseNode) that execute
a query on the EXPLAIN_STREAM, EXPLAIN_OPERATOR tables that extract
information about the execution plan of a statement.

The DBTuneInstances class contains a new utility method that streamlines
the creation of ResultSet mocks. With it, it easy straight-forward to
test DBMS-dependent code.

The DB2OptimizerTest class checks the two new methods in DB2Optimizer.