issue #136 Consistent assignment of update costs
Modifies the constructor of ExplainedSQLStatement so that receives the
updateCost, baseTableUpdateCost and indexUpdateCosts. Then it checks
that the following holds:

getUpdateCost(getUpdatedConfiguration()) + getBaseTableUpdateCost() == getUpdateCost()

This breaks the DB2Optimizer implementation since currently it doesn't
have the ability of obtaining the update costs for each index, only the
aggregated updateCost.

For PGOptimizer this is not a problem although the basetTableUpdateCost
is set to 0.0, which is not actually correct but it's logically
consistent.

For the IBGOptimizer, this doesn't affect it (yet) since the current
version can't handle updates. The same applies to MySQLOptimizer