issue #120 Fixing bug in construction
The configuration that was used to check the usedSet was incorrect. The
code was using the universe instead of the one corresponding to the
newly created node.

Also, this commit moves the analysis phase out of the construction. This
will allow for a better refactoring of the advisor.interaction package.