fixes #145 Updating InteractionBank
Updates the InteractionBank so that it's consistent with the way in
which candidate index sets are passed among the components of the API.

Essentially, a set is transformed into a list and a one-to-one mapping
is obtained between the elements of the internal array-based structures
and the candidate set passed in the constructor. This allows to be
consistent in terms of the position of each index in the internal arrays

The UnionFind structure is not generic, it operates on integers only.