fixes #147 Adding DegreeOfInteractionFinder implementations
DefaultDoiFinder computes the degree of interaction using a regular
optimizer. This is inefficient but useful for testing purposes.

The IBGAnalyzer class was moved to the interactions package and is used
by IBGDoiFinder. This implementation checks the type of prepared
statements that are passed to it and:

 * if an instance of IBGPreparedSQLStatement is given, it retrieves the
   corresponding   IndexBenefitGraph and applies the qINTERACT algorithm
   to it.
 * for any other prepared statement implementation, it calls the
   parent (DefaultDoiFinder).

Test for these classes will be implemented as part of issue #148

dbtune.util.Sets is a utility class used to complement Guava's
com.google.collect.Sets class.