issue #152 Adding Inum-specific execution plan info
Adds an IndexFullTableScan (that extends from Index) that corresponds to
doing a FTS over a base table.

Adds a `TableAccessSlot` class that derives from `Operator`. This is
used to represent a slot in an INUM template plan. It provides
facilities to extract the cost and explore the index associated to it,
which can be a regulare Index object or an IndexFullTableScan. An
InterestingOrder class could potentially be replacing this, although we
need to test in order to determine this

This commit also extends the functionality from the core
dbtune.optimizer package to provide the information that the INUM
extension needs.

In short, `InumPlan` extends from `SQLStatementPlan` class and adds:

 1. `double getInternalCost`
 2. `double plugIntoSlots(Set<Index> atomicConfiguration)`
 3. `List<TableAccessSlot> getAccessSlots`

Constructor InumPlan#InumPlan(SQLStatementPlan) explores the leafs of
the given plan and creates instances of TableAccessSlot that are used to
answer methods 2 and 3. The constructor also extracts the cost of the
internal subplan, to answer 1.

`plugIntoSlots` hasn't been implemented yet.

These methods need to be thoroughly tested via unit and functional
tests.