issue #152 Adding InumPlan functionality
Adds a getFullTableScanIndexInstance method that implements the
singleton-per-table mechanism. The javadoc is also updated

Finishes the implementation of the TableAccessSlot class (javadoc
inclusive)

Adds a copy constructor to Operator (invoked from TableAccessSlot)

Hiddes a node's children in the Tree.Entry class

Adds Index.isCoveredBy method that determines if an index object is
covered by another one. The equals() method was refactored in order to
use the isCoveredBy instead of copy-pasting code

Renames SQLStatementPlan.getReferencedTables() to
SQLStatementPlan.getTables()

Finishes the implementation of InumPlan (including javadoc). The plug()
methods are the ones used to "instantiate" a template plan.