issue #166 Adding an InterestingOrder class
This class derives from Index but nullifies the containment logic that
is imposed by the DatabaseObject class, since we're only interested in
reusing the comparison aspect of an Index object.

FullTableScanIndex now derives from this class.

DerbyInterestingOrdersExtractor now instantiates InterestingOrder
objects instead of Index ones.

Adds a couple of constructors to Index class in order to support this
new derivatives

Adds javadoc for EagerSpaceComputation and InumSpaceComputation