issue #173 Enforcing index-usage by DB2Optimizer
Extends the Optimizer interface to include methods for exposing this new
feature. The AbstractOptimizer implements them in such a way that each
concrete implementation has to check whether or not this is requested.

Adds two methods to the DB2Optimizer class that load optimization
profiles into the OPT_PROFILE table in order to disable FTS on the
indicated tables.

MySQL and Postgres implementations throw exceptions when this new
functionality is invoked.

A MetadataUtils class was added that contains a getTables() method for
obtaining the tables referenced by a collection of indexes

The EagerSpaceComputation class now disables FTS for non-empty
interesting orders in order to obtain the corresponding cost.