issue #178 Fixing the extraction of columns and predicates
The format of the columnNames that DB2 stores on its tables is +Q2.column_name(A) or +Q2.column_name(D) or +Q2.column_name so we had to adapt it for it.

The query to retrieve the predicate list for each operator needed to use LEFT OUTER JOIN instead of INNER to account for operators that don't have entries in the EXPLAIN_PREDICATE table

The parsing of columns in the recommendIndexes method (from the getUsedIndexes method) was being done in the opposite way.