issue #196 Adding tests for InumPlan
InumPlanTest checks the internal cost extraction routine.

InumPlanFunctionalTest calculates the cost of a regular SQLStatementPlan
externally and compares this against the one contained in the
corresponding InumPlan instance. It also checks that the number of slots
corresponds to the number of tables that are queried (for atomic plans)
and that the database objects referenced by the InumPlan are correct.

The above was able to diagnose the SYSIBM issue (see below comment for
DB2Optimizer changes). Another issue that was identified had to do with
how the leafs are reporting the set of database objects used

Tests for the plugging mechanism are needed in order to close this
issue. In particular:

 * plug(Index)
 * plug(Set<Index>)
 * buildQueryForUnseenIndex()

-----

DB2Optimizer now excludes 'GENROW' objects since they're created by the
DB2 optimizer but are not relevant to us.

This commit also adds the modified version of the TPC-H workload used
for the CoPhy paper.

The columnsFetched field was migrated from TableAccessSlot to Operator.

OptimizerVSDelegateFunctionalTest is now using the CandidateGenerator
class. This test will diagnose many issues. Currently, the
ExhaustiveMatchingStrategy is making it fail.