issue #191 Adding changes to matching strategy
Adding an AbstractMatchingStrategy that ensures that all tables
referenced by the statement (that corresponds to the INUM space) contain
at least one index. If the given configuration doesn't contain one, the
FTS index is assigned to it.

Implementations of it should then don't worry about doing this and can
assume that at least one index is included per each table. Also, all the
indexes that don't correspond to any table in the statement are removed
from the matching since they don't matter.

Adds some utility methods to MetadataUtils and a new constructor to
Index