issue #195 Adding TableAccessSlotTest
This checks that a slot is created correctly, given a leaf Operator

This uncovered an important bug in FullTableScanIndex class. In order
for two FTSIndex instances to be compatible, they have to be the same
singleton instance over the same table. This wasn't happening before.

This commit also adds a test to check the covering logic in the Index
class.