issue #209 Instantiating plans from INUM templates
Adds the capability of instantiating a SQLStatementPlan out of an
INUM template (an InumPlan object) by creating Operator objects for each
of the slots in the template.

This enhances the visualization between an INUM plan and the plan
returned by the optimizer and allows to quickly identify the differences
between what INUM is doing/missing.