fixes #173 Enforcing index usage
Fixes a bug in DB2Optimizer.extractColumnsUsedByOperator() that didn't
consider the possibility that the only column referenced by an operator
would be the $RID$, in which case the list of columns is empty and the
InterestingOrder constructor throws an exception.

Adds the correct implementation of setFTSDisabled() to DB2Optimizer, by
writing and loading the XML profile in the way that DB2 expects it. It
also modifies the way the name of a hypothetical index is added to the
ADVISE_INDEX table. Before, only the name was being added but now the
fully qualified name of an index is the value assigned to the
ADVISE_INDEX.NAME column.

Based on this change, the way that objects used by an operator are
parsed changed. The utility method MetadataUtils.find() now looks first
for an index by looking to the fully qualified name. If not found, it
falls back to the old way, i.e. looking only by name.

With this new enforcing mechanism active, InumPlan.instantiate(Index)
now assumes that every index will be used by the optimizer to obtain the
plan for an unseen index.

InumPlan.buildQueryForUnseenIndex() replaces the literals used in
predicates by '?' since the OPTIMIZATION_PROFILE mechanism can still
match queries of this form. In this way we avoid having to escap single
quotes.

DB2Optimizer now passes OptimizerFunctionalTest.