fixes #226 Modifying the way a what-if plan is analyzed
Disables the enforcement of indexes since enforcement they are specific
for a data-access type. An alternative was to do two what-if calls, one
for IXSCAN and another one with LPREFETCH but the costs obtained are
accurate, so we don't need to bother, at least for now.

Adds a INCOMPATIBLE field that is used to avoid returning null values
(since doing so breaks the caching mechanism)

Adds equals and hashCode to a slot and a template.

Tests for most of the above. Still some things should be tested, like
the caching