issue #185 Fixing hashCode() methods
Classes giving Set<> semantics to List<> should take care when defining
the hashCode() method since the order in which elements are stored in a
list influences the value it returns. If a list is still needed, instead
of calling the List.hashCode(), the code should iterate over its
elements and get the sum of the hashcodes, i.e. reimplement it as
specified by the Set<> interface.

The above is true also for the equals() method.