fixes #229 fixes #186; Adding support for join in FROM lists
The Predicate class now can be used to represent join predicates too.

The derby .jar has been updated to reflect the latest changes done to
the Derby source code in order to access JoinNode internal members.

Changes the way that the columns fetched are processed in within
Operator instances, when comparing through equals() and hashCode()

The AbstractInumSpaceComputation now deals with the handling of the
interesting orders returned by the DerbyInterestingOrderExtractor and
leaves implementations to worry about the computation of the space

This commit also adds many tests to check that the interesting orders
extractor functions appropriately