fixes #180 Adding GreedyMatchingStrategy
An alternative to (and faster than) ExhaustiveMatchingStrategy

A test that checks the correctness of the strategy via comparison of
generated plans against those generated by the exhaustive approach.

A new setting is added to EnvironmentProperties and Environment that
allows the selection of the strategy from the configuration file
(config/dbtune.cfg). The optimizer instantiate now calls the matching
factory to instantiate the corresponding class.

Adds also javadoc to the inum package.