issue #190 InterestingOrder won't modify IN_MEMORY_ID
Modifies the InterestingOrder class so that it doesn't increment the
IN_MEMORY_ID counter, otherwise the singleton counter gets incremented
without creating actual Index instances.

This behavior was having as side-effect, in DB2, that the
ADVISE_INDEX.IID column, whose type is SMALLINT would quickly be
assigned with values that are over the range of this datatype and thus
the JDBC driver would throw exceptions

In the near future, the InterestingOrder class will be replaced by a
ColumnOrdering class, which will solve all these type of annoying
issues.