fixes #230 Handling TEMP operator in DB2Optimizer
Transfers the postprocessing tasks from DB2Optimizer to InumUtils since
they're (plus the ones that have been added) are now INUM-specific.
These are now invoked from the InumPlan constructor (renamed to
InumUtils.preprocess()).

One new pre-processing task added is the identification of
TEMPORARY_TABLE_SCAN operators, which correspond to the creation of
temporary tables. These aren't actual base tables and should be ignored
by the INUM framework.

Another issue that this commit solves: the optimizer reuses these TEMP
operators but, from the point of view of the plan-parsing routing, this
re-usage makes it believe that there are distinct base tables being
read. The solution is to parse the plan in two passes. First all the
operator information is obtained, without taking into account the
structure of the plan. In the second phase, the parent-child
relationship is extracted, taking into account that when the same
operator refers to distinct parents, a new operator is created.

DB2OptimizerTest now includes a test for the above.