Miscellaneous fixes
Truncates the name of an InumInterestingOrder to avoid SQLExceptions
being thrown by the DB2Optimizer

The CREATE_TEXT column in the ADVISE_INDEX table for DB2Optimizer is not
assigned since it also causes truncation issues and it's not used by the
what-if mechanism

InumPlan now executes a what-if call on the original query ALWAYS and
disables the query-creation mechanism included before. This enhances the
estimates significantly and avoids getting exceptions from the
underlying driver.