issue #110 Bringing all of WFIT classes from Karl's framework
Brings all the main classes in Karl's SATuning framework, with the
exception of:

 * IBG
 * Connectivity
 * Metadata

The above is already in DBtune and has to be plugged, i.e. a thin
compatibility layer will be added that will take care of mapping between
DBTune's and Karl's entities.

Bringing these classes will allow to have a version of WFIT running
sooner, without having to worry about integration issues. For example:

 * refactor DynamicIndexSet by Set<Index>
 * refactor StaticIndexSet by immutable Set<Index>
 * IndexPartitions by Set<Set<Index>>
 * MinQueue with java.util.Queue
 * ProfiledQuery,AnalyzedQuery by ExplainedSQLStatement
 * ...
 * etc.

All of these are secondary; they can be worked out as time permits