fixes #266 handling of INSERT/DELETE statements in INUM
DB2Optimizer now handles INSERT/DELETE, in the same way that UPDATEs are
handled, i.e. by doing an extra query on the EXPLAIN_STREAM table to
identify the base table being touched.

InumPlan handles DELETE/INSERT statements by ignoring the subtree that
is below the corresponding operator. In other words, for these type of
statements, slots are not created. When a template is instantiated, the
same plan is returned, regardless of the set of indexes sent to the
instantiate() method. When a plug() is invoked, the internal cost is
always returned.

Each test in OptimizerTest was extended in order to execute an INSERT
and a DELETE on each

RefreshFunctionTPCHGeneration was modified in order to make it
single-quote DATE strings

A new small workload was added that contains 4 statements, 2 deletes and
2 inserts that are run by the functional tests