fixes #276 Run WFIT on top of INUM
Adds the option of running the IBGOptimizer on top of the InumOptimizer

Extends the DatabaseSystem.Factory class to consider more than one
"embedded" optimizer. DatabaseSystemTest (and DBTuneInstances) now
ensures the good behavior of this new way of instantiating optimizers.

The Index internal id counter now begins on 1 (to avoid indexing
issues). The IndexTest has been updated to check this.