fixes #298 Fixing test for voting mechanism
The WF values are correctly updated after a positive/negative vote is
seen.

The issue was related to the way the test was written and not with WFIT.
In particular the UPDATEs being sent actually benefit from having an
index, so voting up the index doesn't have any change on the
recommendation because it's actually cheaper to have an index on A.
executed.

The test for WFIT runs OK on a VM but not on a server (eg. aigaion) due
to the different parameters of DB2

-------------------

Also including miscellaneous changes.