fixes #284 fixes #277 Adding MVC functionality
Interface dbtune.viz.Visualizer has been added and abstracts a
visualization.

TotalWorkPlotter and IndexSetPartitionTable have now
RecommendationStatistics members. They take the information that has to
be displayed from these. They also implement the refresh() method
defined in the Visualizer interface.

WorkloadStream instantiates an Actor to which attaches advisors and
visualizers. Every time that a new statement is processed, the actor
comunicates it to every registered advisor and  visualizer.

WFITRecommendationStatistics was merged into RecommendationStatistics