fixes #279 Index-handling
ColumnOrdering#newOrdering parses a string and instantiates the ordering
for an index. DatabaseSystem#newIndex instantiates the index
appropriately (by filling the dbms-dependent member information, like
creation cost and size)

Adds a utility method to MetadataUtils for loading index definitions
from a file.

Database (on the CLI) has newIndex and loadIndexes now

Renamed columns of the one_table workload

Changed the Index#toString method and ExplainedSQLStatement#toString in
order to ease the text-representation of their members