issue #314 negative cost
Some time the cost return by DB2 is inconsistent. The following plan
appeared when evaluating TPCDS first query using optimal indexes.
            │   ├── SORT.MERGE.JOIN(id=5 OUTER cost=2539.47)
            │   │   ├── INDEX.SCAN(id=6 OUTER cost=640633.19)
            │   │   └── FILTER(id=7 INNER cost=12.34)
This fix scan the plan and modify invalid values. If the cost of a child
node is greater than the cost of a parent node, then the cost of the
child node will be modified to (costOfParent-costOfOtherChildern). If
the costs of multiple child are greater than the parent, an exception
will be thrown.