issue #313 add windowOnly parameter
Turn on DATPaperMain.windowOnly so that only one experiment
will be run for each workload using default parameters. The minCost
and maxCost of workload are printed out, so as the cost of each
window. The target is to tune the parameters so that the first
window has a cost close to maxCost and the last window has a cost
close to minCost.