Workload reader abstraction and implementations
A workload reader is an instantiator of SQLStatement objects. Each is
attached to a workload source (sql file, query log, database, etc.).

An ObservableWorkloadReader that facilitates the consumption of sql
statements by keeping an internal watcher thread that periodically
checks for new statements added to a particular source.

QueryLogReader is the common parent for classes that read a workload
from a DBMS' query log.

Tests have been added that check the functionality of these two new
classes.